minimax(bugué)

+++ include/evaluation.h
@@ -8,7 +8,7 @@
 #include <stdbool.h>
 #include "jeu.h"
 
-int maxScore(Jeu* jeu);
+double maxScore(Jeu* jeu);
 double evalFinPartie(Jeu* jeu);
 
 #endif // EVALUATION_H
